Refactor Data Seeding

+++ ASP.NET-MVC/BookShop/BookShop.Mvc/Models/DatabaseContext.cs
@@ -29,14 +29,14 @@ namespace BookShop.Mvc.Models
             //Seed Data
             var books = new List<Book>
             {
-                new Book { Id = 1, Author = "Gojko Adzic", Title = "Bridging the Communication Gap", Price = (decimal)12.20 },
-                new Book { Id = 2, Author = "Gojko Adzic", Title = "Specification By Example", Price = (decimal)15.30 },
-                new Book { Id = 3, Author = "Lisa Crispin and Janet Gregory", Title = "Agile Testing", Price = (decimal)20.20 },
-                new Book { Id = 4, Author = "Mitch Lacey", Title = "The Scrum Field Guide", Price = (decimal)15.31 },
-                new Book { Id = 5, Author = "Martin Fowler", Title = "Refactoring", Price = (decimal)29.55 },
-                new Book { Id = 6, Author = "Esther Derby and Diana Larsen", Title = "Agile Retrospectives", Price = (decimal)16.99 },
-                new Book { Id = 7, Author = "Matt Wynne and Aslak Hellesoy", Title = "The Cucumber Book", Price = (decimal)18.00 },
-                new Book { Id = 8, Author = "David Chelimsky", Title = "The RSpec Book", Price = (decimal)17.50 }
+                new Book { Id = 1, Author = "Gojko Adzic", Title = "Bridging the Communication Gap", Price = 12.20M },
+                new Book { Id = 2, Author = "Gojko Adzic", Title = "Specification By Example", Price = 15.30M },
+                new Book { Id = 3, Author = "Lisa Crispin and Janet Gregory", Title = "Agile Testing", Price = 20.20M },
+                new Book { Id = 4, Author = "Mitch Lacey", Title = "The Scrum Field Guide", Price = 15.31M },
+                new Book { Id = 5, Author = "Martin Fowler", Title = "Refactoring", Price = 29.55M },
+                new Book { Id = 6, Author = "Esther Derby and Diana Larsen", Title = "Agile Retrospectives", Price = 16.99M },
+                new Book { Id = 7, Author = "Matt Wynne and Aslak Hellesoy", Title = "The Cucumber Book", Price = 18.00M },
+                new Book { Id = 8, Author = "David Chelimsky", Title = "The RSpec Book", Price = 17.50M }
             };
 
             foreach (var book in books)
